+++ et-common-libs/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-17172] Quote arbitrary strings instead of func-name:code in profiling.
 ### Changed
 ### Deprecated
 ### Removed