+++ et-common-libs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+- [CI] Fix gitlab-ci default branch
 ### Security
 
 ## [0.16.0] - 2023-04-19