+++ et-common-libs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
+- (CI) Update gitlab-ci-common poointer for Gitlab-CI 15 compatibility
 ### Deprecated
 ### Removed
 ### Fixed