+++ et-common-libs/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - (CI) Update gitlab-ci-common poointer for Gitlab-CI 15 compatibility
 - [SW-17876] Update conanfile.py & CI to gitlab.com
+- Updating the conan dependencies versions to released components
 ### Deprecated
 ### Removed
 ### Fixed