+++ et-common-libs/CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [0.18.0] - 2023-08-03
+### Added
 - [SW-13951] reporting MM state with heartbeat
 ### Changed
 - (CI) Update gitlab-ci-common poointer for Gitlab-CI 15 compatibility